refactor: DRY, DbC, TDD, and mypy quality improvements (#710)
* refactor: DRY constants, add DbC validation, add 134 edge-case tests

- Consolidate ~40 duplicate constants in process_calculators/constants.py
  to import from utils/unit_constants.py (single source of truth)
- Add input validation (DbC) to syngas_compression_calculator,
  acid_gas_dewpoint_calculator, and financial_calculator to catch
  division-by-zero, invalid physics, and out-of-bounds inputs early
- Add 134 parametrized edge-case tests across 4 process calculators:
  syngas compression, acid gas dewpoint, financial model, syngas water


* fix: resolve 70 mypy errors, fix bandit/pre-push hook config

- Fix all 70 mypy errors across 14 process_calculators files:
  - Add float() casts for no-any-return errors
  - Add missing type annotations for no-untyped-def errors
  - Fix attr-defined errors in __init__.py and flow_rate_converter.py
  - Add type: ignore comments for PyQt6 fallback assignments
- Fix bandit pre-push hook: scan only changed files (not entire repo)
- Fix deprecated 'push' stage to 'pre-push' in all hook configs


* fix: remove unused mypy type-ignore comments


* [AutoFix] Iteration 1: Apply fixes ( autoflake)

Automated fixes applied directly to PR #710 branch.
Iteration: 1 of 5
Tools used: autoflake

This commit was created by Jules PR AutoFix workflow.
CI verification loop will check if more fixes are needed.


---------

+++ src/shared/python/upstream_drift_tools/process_calculators/scrubber_calculator.py
@@ -315,7 +315,7 @@ def calculate_gas_viscosity(temperature_k: float, molecular_weight: float) -> fl
 
     # Adjust for molecular weight (heavier gases tend to be more viscous)
     mw_correction = (molecular_weight / MW_AIR_GMOL) ** 0.25
-    return mu * mw_correction
+    return float(mu * mw_correction)
 
 
 def calculate_flooding_velocity(
@@ -423,7 +423,7 @@ def calculate_pressure_drop(
     # Limit to reasonable range
     dp_per_m = min(dp_per_m, ECKERT_MAX_DP_PER_M)  # Max 2 kPa/m (indicates flooding)
 
-    return dp_per_m * packed_height
+    return float(dp_per_m * packed_height)
 
 
 def calculate_ntu_removal(inlet_conc: float, outlet_conc: float) -> float:
@@ -454,7 +454,7 @@ def calculate_ntu_removal(inlet_conc: float, outlet_conc: float) -> float:
 
     # For irreversible absorption (chemical scrubbing)
     # y* ≈ 0, so NTU = ln(y_in/y_out)
-    return np.log(inlet_conc / outlet_conc)
+    return float(np.log(inlet_conc / outlet_conc))
 
 
 def calculate_htu(
@@ -504,7 +504,7 @@ def calculate_htu(
     htu = packing.ch / (kla_per_s * packing.specific_surface_area * l_over_g**packing.n)
 
     # Clamp to reasonable range (0.1 to 3 m)
-    return max(HTU_MIN, min(HTU_MAX, htu))
+    return float(max(HTU_MIN, min(HTU_MAX, htu)))
 
 
 def calculate_required_packed_height(
